Error Checks and Bug Fixes
Added checks to __init__.py and export.py to pause export if an error is encountered due to a missing material on an object or missing texture on an object that is attempting to use texture GBI commands. This also drops an error message in the .c file, blender console window, and a pop up message.